refactor profile pages
