+++ src/app/profile/change-password/change-password-form.tsx
@@ -1,19 +1,13 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import Link from 'next/link';
-import { useAction } from 'next-safe-action/hooks';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { toast } from 'sonner';
-import type { GetProfileResponse } from '@/app/profile/profile-types';
-import { changePasswordAction } from '@/app/profile/actions/change-password-action';
 import { Form } from '@/components/ui/form';
 import { Button } from '@/components/ui/button';
 import { LoadingButton } from '@/components/loading-button';
 import { TextFormField } from '@/components/text-form-field';
-import { ServerActionErrorMessage } from '@/components/server-action-error';
 
 const schema = z
     .object({
@@ -33,15 +27,17 @@ const schema = z
         path: ['confirmNewPassword'],
     });
 
-type ChangePasswordFormValues = z.infer<typeof schema>;
+export type ChangePasswordFormValues = z.infer<typeof schema>;
 
 type ChangePasswordFormProps = {
-    profile: GetProfileResponse;
+    loading?: boolean;
+    onSubmit: (data: ChangePasswordFormValues) => void;
 };
 
-export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
-    const router = useRouter();
-
+export function ChangePasswordForm({
+    loading,
+    onSubmit,
+}: ChangePasswordFormProps) {
     const form = useForm<ChangePasswordFormValues>({
         resolver: zodResolver(schema),
         defaultValues: {
@@ -50,23 +46,6 @@ export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
             confirmNewPassword: '',
         },
     });
-
-    const { execute, isPending } = useAction(changePasswordAction, {
-        onSuccess: () => {
-            toast.success('Your password has been changed.');
-            router.push('/profile');
-        },
-        onError: ({ error }) => {
-            toast.error(<ServerActionErrorMessage result={error} />);
-        },
-    });
-
-    function onSubmit(data: ChangePasswordFormValues) {
-        execute({
-            ...data,
-            version: profile.version,
-        });
-    }
 
     return (
         <Form {...form}>
@@ -83,7 +62,7 @@ export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
                     maxLength={50}
                     autoComplete="current-password"
                     required
-                    disabled={isPending}
+                    disabled={loading}
                 />
 
                 <div className="flex flex-col gap-6 sm:flex-row">
@@ -95,7 +74,7 @@ export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
                         maxLength={50}
                         autoComplete="new-password"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                     <TextFormField
@@ -106,7 +85,7 @@ export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
                         maxLength={50}
                         autoComplete="new-password"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                 </div>
@@ -116,7 +95,7 @@ export function ChangePasswordForm({ profile }: ChangePasswordFormProps) {
                         <Link href="/profile">Cancel</Link>
                     </Button>
 
-                    <LoadingButton type="submit" loading={isPending}>
+                    <LoadingButton type="submit" loading={loading}>
                         Submit
                     </LoadingButton>
                 </div>
