+++ src/app/profile/update-profile/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import { getProfileAction } from '@/app/profile/actions/get-profile-action';
-import { UpdateProfileForm } from '@/app/profile/update-profile/update-profile-form';
+import { UpdateProfile } from '@/app/profile/update-profile/update-profile';
 import {
     isServerActionSuccess,
     ServerActionError,
@@ -17,20 +17,5 @@ export default async function UpdateProfilePage() {
         return <ServerActionError result={result} />;
     }
 
-    const profile = result.data;
-
-    return (
-        <main className="mx-auto max-w-screen-sm space-y-6 p-6">
-            <h1 className="scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-5xl">
-                Update Profile
-            </h1>
-
-            <p className="leading-7">
-                Update your personal details below. Your email address is used
-                to login to your account.
-            </p>
-
-            <UpdateProfileForm profile={profile} />
-        </main>
-    );
+    return <UpdateProfile profile={result.data} />;
 }
