+++ src/app/profile/update-profile/update-profile-form.tsx
@@ -1,20 +1,14 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import Link from 'next/link';
-import { useAction } from 'next-safe-action/hooks';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { toast } from 'sonner';
-import type { GetProfileResponse } from '@/app/profile/profile-types';
-import { updateProfileAction } from '@/app/profile/actions/update-profile-action';
 import { Form } from '@/components/ui/form';
 import { Button } from '@/components/ui/button';
 import { LoadingButton } from '@/components/loading-button';
 import { DateFormField } from '@/components/date-form-field';
 import { TextFormField } from '@/components/text-form-field';
-import { ServerActionErrorMessage } from '@/components/server-action-error';
 import { isInPast } from '@/lib/dates';
 
 const schema = z.object({
@@ -37,36 +31,23 @@ const schema = z.object({
         .refine(isInPast, { message: 'Date Of Birth must be in the past' }),
 });
 
-type UpdateProfileFormValues = z.infer<typeof schema>;
+export type UpdateProfileFormValues = z.infer<typeof schema>;
 
 type UpdateProfileFormProps = {
-    profile: GetProfileResponse;
+    values: UpdateProfileFormValues;
+    loading?: boolean;
+    onSubmit: (data: UpdateProfileFormValues) => void;
 };
 
-export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
-    const router = useRouter();
-
+export function UpdateProfileForm({
+    values,
+    loading,
+    onSubmit,
+}: UpdateProfileFormProps) {
     const form = useForm<UpdateProfileFormValues>({
         resolver: zodResolver(schema),
-        values: profile,
+        values,
     });
-
-    const { execute, isPending } = useAction(updateProfileAction, {
-        onSuccess: () => {
-            toast.success('Your profile has been updated.');
-            router.push('/profile');
-        },
-        onError: ({ error }) => {
-            toast.error(<ServerActionErrorMessage result={error} />);
-        },
-    });
-
-    function onSubmit(data: UpdateProfileFormValues) {
-        execute({
-            ...data,
-            version: profile.version,
-        });
-    }
 
     return (
         <Form {...form}>
@@ -83,7 +64,7 @@ export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
                     maxLength={254}
                     autoComplete="username"
                     required
-                    disabled={isPending}
+                    disabled={loading}
                 />
 
                 <div className="flex flex-col gap-6 sm:flex-row">
@@ -94,7 +75,7 @@ export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
                         maxLength={50}
                         autoComplete="given-name"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                     <TextFormField
@@ -104,7 +85,7 @@ export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
                         maxLength={50}
                         autoComplete="family-name"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                 </div>
@@ -115,7 +96,7 @@ export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
                     label="Date Of Birth"
                     autoComplete={['bday-day', 'bday-month', 'bday-year']}
                     required
-                    disabled={isPending}
+                    disabled={loading}
                 />
 
                 <div className="flex flex-col-reverse justify-end gap-2 sm:flex-row">
@@ -123,7 +104,7 @@ export function UpdateProfileForm({ profile }: UpdateProfileFormProps) {
                         <Link href="/profile">Cancel</Link>
                     </Button>
 
-                    <LoadingButton type="submit" loading={isPending}>
+                    <LoadingButton type="submit" loading={loading}>
                         Submit
                     </LoadingButton>
                 </div>
